Create Exceptions.cs and add ContainerHttpException
